issue #33 削除時に確認ダイアログを表示

+++ src/view/page/UpdateSchedulePage.tsx
@@ -66,6 +66,10 @@ export const UpdateSchedulePage: FC = () => {
   );
 
   const handleDelete = useCallback(() => {
+    const ok = confirm('削除します');
+    if (!ok) {
+      return;
+    }
     deleteTaskSchedule(id);
     navigate('/schedules');
   }, [deleteTaskSchedule, id, navigate]);
